new file for movement for norm
