+++ src/init/mlx.c
@@ -6,61 +6,11 @@
 /*   By: mdekker <mdekker@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/01/17 15:06:14 by mdekker       #+#    #+#                 */
-/*   Updated: 2024/02/10 03:38:52 by mdekker       ########   odam.nl         */
+/*   Updated: 2024/02/14 01:22:08 by maxvalk       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <cub3d.h>
-
-bool	is_floor(t_data *data, double x, double y)
-{
-	int	mapx;
-	int	mapy;
-
-	mapx = (int)x;
-	mapy = (int)y;
-	if (data->map.array[mapy][mapx] == FLOOR
-		|| data->map.array[mapy][mapx] == PLAYER
-		|| data->map.array[mapy][mapx] == OPEN_DOOR
-		|| data->map.array[mapy][mapx] == SPRITE)
-		return (true);
-	return (false);
-}
-
-void	set_new_pos(t_data *data, t_player player, t_info_types dir,
-		double incr)
-{
-	double	newx;
-	double	newy;
-
-	newx = player.x;
-	newy = player.y;
-	if (dir == S)
-	{
-		newx = player.x + cos(player.dir) * incr;
-		newy = player.y + sin(player.dir) * incr;
-	}
-	else if (dir == N)
-	{
-		newx = player.x - cos(player.dir) * incr;
-		newy = player.y - sin(player.dir) * incr;
-	}
-	else if (dir == W)
-	{
-		newx = player.x + sin(player.dir) * incr;
-		newy = player.y - cos(player.dir) * incr;
-	}
-	else if (dir == E)
-	{
-		newx = player.x - sin(player.dir) * incr;
-		newy = player.y + cos(player.dir) * incr;
-	}
-	if (is_floor(data, newx, newy))
-	{
-		data->player.x = newx;
-		data->player.y = newy;
-	}
-}
 
 void	draw_sprite(t_data *data)
 {
